+++ ll/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.6.0] - 2021-02-08
 ### Added
 - Implement `fmt::Display` for `SocketInterrupt`.
 - Implement `defmt::Format` for `Ipv4Addr` and `SocketAddrV4`.
@@ -80,7 +80,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2021-01-02
 - Initial release
 
-[Unreleased]: https://github.com/newAM/w5500-ll-rs/compare/v0.5.1...HEAD
+[Unreleased]: https://github.com/newAM/w5500-ll-rs/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/newAM/w5500-ll-rs/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/newAM/w5500-ll-rs/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/newAM/w5500-ll-rs/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/newAM/w5500-ll-rs/compare/v0.3.0...v0.4.0